# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Academic Publishing Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /submissions:
    get:
      operationId: getSubmissions
      tags: [submissions]
      summary: List submissions (filtered by role)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionListResponse"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Create a new submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionRequest"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get submission details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /submissions/{id}/status:
    patch:
      operationId: updateSubmissionStatus
      tags: [submissions]
      summary: Update submission status (editor/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/assign:
    post:
      operationId: assignReviewer
      tags: [submissions]
      summary: Assign a reviewer to a submission (editor/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignReviewerRequest"
      responses:
        "200":
          description: Reviewer assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /submissions/{id}/upload:
    post:
      operationId: uploadDocument
      tags: [submissions]
      summary: Upload document for submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                docType:
                  type: string
      responses:
        "200":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
  /admin/audit-logs:
    get:
      operationId: getAuditLogs
      tags: [admin]
      summary: Get audit logs (admin only)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
...
  /admin/email-templates/{id}:
    patch:
      operationId: updateEmailTemplate
      tags: [admin]
      summary: Update email template (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmailTemplateRequest"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    RegisterRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
        departmentId:
          type: integer
        scientificDegree:
          type: string
          enum: [PhD, DSc, none]
        position:
          type: string
          enum: [teacher, senior_teacher, associate_professor, professor]
      required: [fullName, email, password, departmentId, scientificDegree, position]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [author, editor, reviewer, admin]
        departmentId:
          type: integer
        departmentName:
          type: string
        scientificDegree:
          type: string
        position:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, fullName, email, role]
...
    Submission:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        abstract:
          type: string
        keywords:
          type: array
          items:
            type: string
        language:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        scientificDirection:
          type: string
        literatureType:
          type: string
          enum: [darslik, oquv_qollanma, monografiya, oquv_uslubiy_qollanma, uslubiy_korsatma]
        status:
          type: string
          enum: [submitted, under_review, revision_required, accepted, rejected, published]
        authorId:
          type: integer
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, literatureType, status, authorId, createdAt]
    SubmissionDetail:
      allOf:
        - $ref: "#/components/schemas/Submission"
        - type: object
          properties:
            documents:
              type: array
              items:
                $ref: "#/components/schemas/Document"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            editorNotes:
              type: string
    SubmissionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Submission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [items, total]
    CreateSubmissionRequest:
      type: object
      properties:
        title:
          type: string
        abstract:
          type: string
        keywords:
          type: array
          items:
            type: string
        language:
          type: string
        departmentId:
          type: integer
        scientificDirection:
          type: string
        literatureType:
          type: string
          enum: [darslik, oquv_qollanma, monografiya, oquv_uslubiy_qollanma, uslubiy_korsatma]
      required: [title, abstract, departmentId, scientificDirection, literatureType]
    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [submitted, under_review, revision_required, accepted, rejected, published]
        notes:
          type: string
      required: [status]
    AssignReviewerRequest:
      type: object
      properties:
        reviewerId:
          type: integer
      required: [reviewerId]
    Document:
      type: object
      properties:
        id:
          type: integer
        submissionId:
          type: integer
        docType:
          type: string
          enum: [internal_review, external_review, plagiarism_report, curriculum, syllabus, main_document]
        fileName:
          type: string
        fileSize:
          type: integer
        filePath:
          type: string
        uploadedAt:
          type: string
          format: date-time
      required: [id, submissionId, docType, fileName]
    Review:
      type: object
      properties:
        id:
          type: integer
        submissionId:
          type: integer
        submissionTitle:
          type: string
        reviewerId:
          type: integer
        reviewerName:
          type: string
        scientificSignificance:
          type: integer
          minimum: 1
          maximum: 10
        methodology:
          type: integer
          minimum: 1
          maximum: 10
        structureClarity:
          type: integer
          minimum: 1
          maximum: 10
        originality:
          type: integer
          minimum: 1
          maximum: 10
        commentsForAuthor:
          type: string
        commentsForEditor:
          type: string
        verdict:
          type: string
          enum: [accept, minor_revision, major_revision, reject]
        status:
          type: string
          enum: [pending, submitted]
        assignedAt:
          type: string
          format: date-time
        submittedAt:
          type: string
          format: date-time
      required: [id, submissionId, reviewerId, status]
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userEmail:
          type: string
        userRole:
          type: string
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        detail:
          type: string
        ipAddress:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, action, createdAt]
    AuditLogListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
      required: [items, total]
    EmailTemplate:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, key, name, subject, body, isActive]
    UpdateEmailTemplateRequest:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        isActive:
          type: boolean