# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academic Publishing Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface AuditLog {
  id: number;
  userId?: number;
  userEmail?: string;
  userRole?: string;
  action: string;
  entityType?: string;
  entityId?: string;
  detail?: string;
  ipAddress?: string;
  createdAt: Date;
}