# lib/api-zod/src/generated/types/auditLogListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academic Publishing Portal API
 * OpenAPI spec version: 1.0.0
 */
import type { AuditLog } from "./auditLog";

export interface AuditLogListResponse {
  items: AuditLog[];
  total: number;
}