# lib/api-zod/src/generated/types/emailTemplate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academic Publishing Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface EmailTemplate {
  id: number;
  key: string;
  name: string;
  subject: string;
  body: string;
  isActive: boolean;
  createdAt?: Date;
  updatedAt?: Date;
}