# lib/api-zod/src/generated/types/updateEmailTemplateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academic Publishing Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateEmailTemplateRequest {
  subject?: string;
  body?: string;
  isActive?: boolean;
}